spring-beginning/src/main/java/com/example/spring/repository/MemoryMemberRepository.java
package com.example.spring.repository;

import com.example.spring.domain.Member;
import org.springframework.stereotype.Repository;

import java.util.*;


// @Repository  // componet scan (spring Bean)
public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;  // 동시성 문제로 AtomicLong 으로 해야한다.

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByUsername(String username) {
        return store.values().stream()
                .filter(member -> member.getUsername().equals(username))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
